requirements.txt
hyperliquid-python-sdk>=0.10.1
pandas>=1.5.0
numpy>=1.23.0
colorama>=0.4.6
schedule>=1.1.0
termcolor>=2.1.0